Moved to turso/sqlite only as db & Fixed db transactions /w query logging & Minor UI improvement

+++ lib/routes/replication.server.ts
@@ -1,5 +1,5 @@
 import { Bucket } from "../bucket";
-import { mysqlGenerateTables, sqliteGenerateBackup } from "../routes/schema.server";
+import { sqliteGenerateBackup } from "../routes/schema.server";
 import { executeCommands } from "../utils.cli";
 import { asyncQueue, deepCopy } from "../utils.client";
 import { MIMETypeMap, execTryCatch } from "../utils.server";
@@ -7,11 +7,22 @@ import { ReplicationRoutes } from "./replication.client";
 
 const serverRoutes = deepCopy(ReplicationRoutes); // Copy the routes object to split it into client and server routes
 
-async function productionBucketReplication() {
-	// This is a dev only function to replicate the dev bucket from the prod bucket
-	// Wipe the dev bucket
-	const devBucketList = await Bucket.listDev();
+const getCurrentFormattedDate = () => {
+	const d = new Date();
+	return `${d.getFullYear().toString().slice(-2)}-${(d.getMonth() + 1).toString().padStart(2, "0")}-${d.getDate().toString().padStart(2, "0")}`;
+};
 
+async function productionBucketReplication() {
+	const { DEV_BUCKET_LOCATION, S3_BUCKET_NAME } = await import.meta.env;
+	if (!DEV_BUCKET_LOCATION) throw Error("Missing environment variables");
+	const fs = (await eval('import("fs/promises")')) as typeof import("fs/promises");
+	// This is a dev only function to replicate the dev bucket from the prod bucket
+	// Locally and in cloudflare
+
+	const devBucketList = await Bucket.listDev();
+	const BUCKET_DATE = getCurrentFormattedDate();
+
+	// Wipe the dev bucket
 	await asyncQueue(devBucketList.map((file) => {
 		return () => Bucket.deleteDev(file);
 	}), {
@@ -19,18 +30,21 @@ async function productionBucketReplication() {
 	});
 	console.log("Dev bucket wiped");
 
-	const prodBucketList = await Bucket.listDev("byzantini-bucket");
+	const prodBucketList = await Bucket.listDev(S3_BUCKET_NAME);
 	let totalReplicatedFiles = 0;
 	await asyncQueue(prodBucketList.map((fileName) => {
 		return async () => {
 			const fileType = fileName.split(".").at(-1);
 			if (!fileType) throw Error("File type not found");
 
-			const file = await Bucket.getDev(fileName, "byzantini-bucket");
+			const file = await Bucket.getDev(fileName, S3_BUCKET_NAME);
 			if (!file) throw Error("File not found");
 
 			await Bucket.putDev(file, fileName, MIMETypeMap[fileType] || "application/octet-stream");
+			await fs.mkdir(`${DEV_BUCKET_LOCATION}/${BUCKET_DATE}/${fileName.split("/").slice(0, -1).join("/")}`, { recursive: true });
+			await fs.writeFile(`${DEV_BUCKET_LOCATION}/${BUCKET_DATE}/${fileName}`, Buffer.from(file), { encoding: "utf-8" });
 			totalReplicatedFiles++;
+
 		};
 	}), {
 		maxJobs: 10,
@@ -41,68 +55,41 @@ async function productionBucketReplication() {
 	else console.log("Prod bucket replicated to dev bucket successfully");
 }
 
-// You'll need to authenticate with `pscale auth login` before running this function
-async function productionDatabaseReplication() {
-	const { REPLICA_DB_PWD, REPLICA_DB_NAME } = await import.meta.env;
-	if (!REPLICA_DB_PWD || !REPLICA_DB_NAME) throw Error("Missing environment variables");
+
+async function productionDatabaseReplication(force = false) {
+	const { BACKUP_SNAPSHOT_LOCATION, DEV_SNAPSHOT_LOCATION, PROJECT_ABSOLUTE_PATH } = await import.meta.env;
+	if (!BACKUP_SNAPSHOT_LOCATION || !DEV_SNAPSHOT_LOCATION || !PROJECT_ABSOLUTE_PATH) throw Error("Missing environment variables");
 	const fs = (await eval('import("fs/promises")')) as typeof import("fs/promises");
-	const d = new Date();
-	const SNAPSHOT_DATE = `${d.getFullYear().toString().slice(-2)}-${(d.getMonth() + 1).toString().padStart(2, "0")}-${d.getDate().toString().padStart(2, "0")}`;
+	const SNAPSHOT_DATE = getCurrentFormattedDate();
 
 	let fileBackup;
-	try {
-		fileBackup = await fs.readFile(`C:/Users/poupa/Projects/Javascript/astro/byzantini-website/dbSnapshots/sqlite/snap-${SNAPSHOT_DATE}.sql`, {
-			encoding: "utf-8"
-		});
-	} catch (error) {
-
+	if (!force) {
+		try {
+			fileBackup = await fs.readFile(`${BACKUP_SNAPSHOT_LOCATION}/snap-${SNAPSHOT_DATE}.sql`, {
+				encoding: "utf-8"
+			});
+		} catch (error) {
+			console.log("No backup found for today, generating a new one");
+		}
 	}
 	const sqliteBackup = fileBackup || await sqliteGenerateBackup();
 
 	// Store sqlite file locally
-	await fs.writeFile(`C:/Users/poupa/Projects/Javascript/astro/byzantini-website/dbSnapshots/sqlite/snap-${SNAPSHOT_DATE}.sql`, sqliteBackup, {
+	await fs.writeFile(`${BACKUP_SNAPSHOT_LOCATION}/snap-${SNAPSHOT_DATE}.sql`, sqliteBackup, {
 		encoding: "utf-8"
 	});
 
-	let mysqlTables = await mysqlGenerateTables();
-	mysqlTables = mysqlTables.map((createStr) => {
-		const tableName = createStr.match(/`(\w+)`/)?.[1] || "";
-		return {
-			tableName,
-			createStr: `DROP TABLE IF EXISTS \`${tableName}\`;\n` + createStr,
-			inserts: []
-		};
-	});
-
-	sqliteBackup.split("\n").forEach(line => {
-		if (line.startsWith("INSERT")) {
-			const tableName = line.match(/INTO (\w+) \(/)?.[1];
-			if (!tableName) return;
-			const table = mysqlTables.find(table => table.tableName === tableName);
-			if (!table) return;
-			table.inserts.push(line);
-		}
-	});
-
-
-	const mysqlBackup = mysqlTables.map(table =>
-		`${table.createStr};\n${table.inserts.join("\n")}`
-	).join("\n");
-
-	await fs.writeFile(`C:/Users/poupa/Projects/Javascript/astro/byzantini-website/dbSnapshots/mysql/snap-${SNAPSHOT_DATE}.sql`, mysqlBackup, {
+	await fs.writeFile(DEV_SNAPSHOT_LOCATION, sqliteBackup, {
 		encoding: "utf-8"
 	});
 
-	const comms = [
-		`$OutputEncoding = [Console]::OutputEncoding = [Text.Encoding]::UTF8`,
-		`mysql --defaults-file='C:\\ProgramData\\MySQL\\MySQL Server 8.0\\my.ini' ${REPLICA_DB_NAME}  -u root  -p${REPLICA_DB_PWD} -e 'SOURCE C:/Users/poupa/Projects/Javascript/astro/byzantini-website/dbSnapshots/mysql/snap-${SNAPSHOT_DATE}.sql ;'`,
-	];
+	await executeCommands([
+		`cd ${PROJECT_ABSOLUTE_PATH}/dbSnapshots`,
+		"rm -f latest.db",
+		"sqlite3 latest.db < dev-snapshot.sql",
+	]);
 
-	try {
-		await executeCommands(comms);
-	} catch (error) {
-		console.error(error);
-	}
+	console.log("Database replicated successfully");
 }
 
 
@@ -115,7 +102,7 @@ serverRoutes.replication.func = ({ ctx, slug }) => {
 		// support for the @aws-sdk/client-s3 package or the child_process api
 		// so it would throw an error regardless
 		const { service } = slug;
-		if (service === "database") {
+		if (service === "database" || service === "database-force") {
 			await productionDatabaseReplication();
 			return "Database replicated";
 		}
@@ -124,7 +111,7 @@ serverRoutes.replication.func = ({ ctx, slug }) => {
 			return "Bucket replicated";
 		}
 		else if (service === "both") {
-			await Promise.all([productionDatabaseReplication(), productionBucketReplication()]);
+			await Promise.all([productionDatabaseReplication(true), productionBucketReplication()]);
 			return "Database and bucket replicated";
 		} else throw Error("Invalid service");
 	});
